requirements.txt
streamlit>=1.24.0
anthropic>=0.19.1
python-dotenv>=1.0.0
httpx>=0.24.1 